feat: complete exercise 6.11
Add a simple function to ticker.py that filters rows based on if each
row['name'] is contained in the second argument. The second argument can
be a portfolio from report, allowing a stream of stock prices to be
filtered to those of interest based on their inclusion in a portfolio object.